Unify the scripts into one
Replace hardcoded configuration variables with command-line options.

Switch from optparse to argparse.

Fix handling of --quiet (it used to be completely ignored, now it'll
reduce verbosity by one).

Improve logging in get_published_binaries() (so I don't have to wonder
why it's listing the same thing twice, because it's not the same thing
when the architecture tag differs!).